CheckMessageHistoryUseCase

+++ GameBear/Messages/ICreateNewGameData.cs
@@ -4,7 +4,7 @@ namespace DealerBear.Messages
     {
         string SessionID { get; }
         string MessageID { get; }
-        float Seed { get; }
+        int Seed { get; }
         int PackVersionNumber { get; }
         string CurrentCard { get; set; }
     }
